+++ src/main/java/com/neu/his/dto/DrugPrescriptionDetailDTO.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class DrugPrescriptionDetailDTO {
-    private Short drugId;//药品ID
+    private Integer drugId;//药品ID
     private String drugPreDetailUsage;//药品用法
     private String drugPreDetailAmount;//药品用量
     private String drugPreDetailFreq;//药品频次
